ACK sync

+++ CC-TP2/DataTransferPacket.java
@@ -17,7 +17,7 @@ public class DataTransferPacket {
         this.data = data;
     }
 
-    public short getNumBloco() {
+    public int getNumBloco() {
         return this.numeroBloco;
     }
